automate/pyproject.toml
[tool.poetry]
name = "automate"
version = "0.1.0"
description = ""
authors = ["ENDERZOMBI102 <enderzombi102.end@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
types-PyYAML = "^6.0.10"
CairoSVG = "^2.5.2"
typing-extensions = "^4.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
